[refactoring/daengle-9] 미용 견적서 관련 Swagger 설명 추가 패키지 구조 분리 리팩토링
