+++ src/main/java/ddog/daengleserver/application/GroomingEstimateService.java
@@ -6,12 +6,12 @@ import ddog.daengleserver.application.repository.UserRepository;
 import ddog.daengleserver.domain.Groomer;
 import ddog.daengleserver.domain.User;
 import ddog.daengleserver.domain.estimate.GroomingEstimate;
-import ddog.daengleserver.presentation.dto.request.GroomerGroomingEstimateReq;
-import ddog.daengleserver.presentation.dto.request.UserDesignationGroomingEstimateReq;
-import ddog.daengleserver.presentation.dto.request.UserGeneralGroomingEstimateReq;
-import ddog.daengleserver.presentation.dto.response.UserGroomingEstimateDetails;
-import ddog.daengleserver.presentation.dto.response.UserGroomingEstimateInfo;
-import ddog.daengleserver.presentation.dto.response.UserAndPetsInfo;
+import ddog.daengleserver.presentation.estimate.dto.request.GroomerGroomingEstimateReq;
+import ddog.daengleserver.presentation.estimate.dto.request.UserDesignationGroomingEstimateReq;
+import ddog.daengleserver.presentation.estimate.dto.request.UserGeneralGroomingEstimateReq;
+import ddog.daengleserver.presentation.estimate.dto.response.UserGroomingEstimateDetails;
+import ddog.daengleserver.presentation.estimate.dto.response.UserGroomingEstimateInfo;
+import ddog.daengleserver.presentation.estimate.dto.response.UserAndPetsInfo;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
