+++ src/main/java/ddog/daengleserver/presentation/estimate/dto/response/UserGroomingEstimateDetails.java
@@ -1,28 +1,30 @@
-package ddog.daengleserver.presentation.dto.request;
+package ddog.daengleserver.presentation.estimate.dto.response;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import ddog.daengleserver.domain.Weight;
+import lombok.Builder;
 import lombok.Getter;
 
 import java.time.LocalDateTime;
 
 @Getter
-public class UserGeneralGroomingEstimateReq {
+@Builder
+public class UserGroomingEstimateDetails {
 
-    private Long userId;
-    private Long petId;
+    private Long groomingEstimateId;
     private String userImage;
-    private String petImage;
     private String nickname;
-    private String petSignificant;
-    private int petBirth;
-    private Weight petWeight;
-    private String petName;
+    private String address;
 
     @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd HH:mm:ss", timezone = "Asia/Seoul")
     private LocalDateTime reservedDate;
+
+    private String petImage;
+    private int petBirth;
+    private Weight petWeight;
+    private String petSignificant;
+    private String petName;
     private String desiredStyle;
     private String requirements;
-    private String address;
 
 }
